Re-use more logic and cosmetics/cleanup
